Навёл порядок в файловой структуре
