--- project/src/components/film-page/film-tab-details/type.ts
@@ -1,5 +1,0 @@
-import type { FilmProps } from '../../../types/film';
-
-export type FilmTabDetailsProps = {
-  film: FilmProps,
-}
