--- project/src/components/film-page/film-tab-reviews/type.ts
@@ -1,5 +1,0 @@
-import type { ReviewProps } from '../../../types/review';
-
-export type FilmTabReviewsProps = {
-  reviews: ReviewProps[],
-}
