--- project/src/components/film-page/film-tabs/type.ts
@@ -1,8 +1,0 @@
-import type { FilmProps } from '../../../types/film';
-import type { ReviewProps } from '../../../types/review';
-
-export type FilmTabsProps = {
-  id: number,
-  film: FilmProps,
-  reviews: ReviewProps[],
-}
