--- project/src/components/main-page/genre-list/type.ts
@@ -1,6 +1,0 @@
-import { FilmProps } from '../../../types/film';
-
-export type GenreListProps = {
-  films: FilmProps[],
-  resetShowSize: () => void,
-}
