--- project/src/components/main-page/main/type.ts
@@ -1,5 +1,0 @@
-import type { FilmProps } from '../../../types/film';
-
-export type MainPageProps = {
-  films: FilmProps[],
-}
