--- project/src/components/my-list/type.ts
@@ -1,5 +1,0 @@
-import type { FilmProps } from '../../types/film';
-
-export type MyListProps = {
-  films: FilmProps[],
-}
