+++ project/src/components/pages/film-page/film-tab-details/film-tab-details.tsx
