+++ project/src/components/pages/film-page/film-tab-details/type.ts
@@ -1,0 +1,5 @@
+import type { FilmProps } from '../../../../types/film';
+
+export type FilmTabDetailsProps = {
+  film: FilmProps,
+}
