+++ project/src/components/pages/film-page/film-tab-reviews/type.ts
@@ -1,0 +1,5 @@
+import type { ReviewProps } from '../../../../types/review';
+
+export type FilmTabReviewsProps = {
+  reviews: ReviewProps[],
+}
