+++ project/src/components/pages/film-page/film-tabs/film-tabs.tsx
