+++ project/src/components/pages/film-page/film-tabs/type.ts
@@ -1,0 +1,8 @@
+import type { FilmProps } from '../../../../types/film';
+import type { ReviewProps } from '../../../../types/review';
+
+export type FilmTabsProps = {
+  id: number,
+  film: FilmProps,
+  reviews: ReviewProps[],
+}
