+++ project/src/components/pages/main-page/genre-list/type.ts
@@ -1,0 +1,6 @@
+import { FilmProps } from '../../../../types/film';
+
+export type GenreListProps = {
+  films: FilmProps[],
+  resetShowSize: () => void,
+}
