+++ project/src/components/pages/main-page/main/type.ts
@@ -1,0 +1,5 @@
+import type { FilmProps } from '../../../../types/film';
+
+export type MainPageProps = {
+  films: FilmProps[],
+}
