+++ project/src/components/pages/main-page/show-more/show-more.tsx
