+++ project/src/components/pages/main-page/show-more/type.ts
