+++ project/src/components/pages/my-list-page/my-list.tsx
@@ -1,0 +1,55 @@
+import FilmList from '../../film-list/film-list';
+import { MyListProps } from './type';
+
+export default function MyList({films}: MyListProps): JSX.Element {
+  const favoriteFilms = films.filter((film) => film.isFavorite);
+
+  return (
+    <div className="user-page">
+      <header className="page-header user-page__head">
+        <div className="logo">
+          <a href="main.html" className="logo__link">
+            <span className="logo__letter logo__letter--1">W</span>
+            <span className="logo__letter logo__letter--2">T</span>
+            <span className="logo__letter logo__letter--3">W</span>
+          </a>
+        </div>
+
+        <h1 className="page-title user-page__title">My list</h1>
+
+        <ul className="user-block">
+          <li className="user-block__item">
+            <div className="user-block__avatar">
+              <img src="img/avatar.jpg" alt="User avatar" width="63" height="63" />
+            </div>
+          </li>
+          <li className="user-block__item">
+            <a href="/" className="user-block__link">Sign out</a>
+          </li>
+        </ul>
+      </header>
+
+      <section className="catalog">
+        <h2 className="catalog__title visually-hidden">Catalog</h2>
+
+        <div className="catalog__films-list">
+          <FilmList films={favoriteFilms} />
+        </div>
+      </section>
+
+      <footer className="page-footer">
+        <div className="logo">
+          <a href="main.html" className="logo__link logo__link--light">
+            <span className="logo__letter logo__letter--1">W</span>
+            <span className="logo__letter logo__letter--2">T</span>
+            <span className="logo__letter logo__letter--3">W</span>
+          </a>
+        </div>
+
+        <div className="copyright">
+          <p>© 2019 What to watch Ltd.</p>
+        </div>
+      </footer>
+    </div>
+  );
+}
