+++ project/src/components/pages/my-list-page/type.ts
@@ -1,0 +1,5 @@
+import type { FilmProps } from '../../../types/film';
+
+export type MyListProps = {
+  films: FilmProps[],
+}
