+++ project/src/components/pages/player-page/type.ts
@@ -1,0 +1,5 @@
+import { FilmProps } from '../../../types/film';
+
+export type PlayerProps = {
+  films: FilmProps[],
+}
