--- project/src/components/player/type.ts
@@ -1,5 +1,0 @@
-import { FilmProps } from '../../types/film';
-
-export type PlayerProps = {
-  films: FilmProps[],
-}
